requirements.txt
flask==3.0.0

PyYAML~=6.0.1
WTForms~=3.0.1
googletrans~=3.0.0

python-dateutil~=2.8.2